M:modify price from double to Decimal

+++ Tilli/Model/TransactionModel.swift
@@ -10,7 +10,7 @@ struct TransactionModel: Identifiable, Codable, Hashable {
     var id = UUID()
     var sessionId: UUID
     var items: [SummaryItemModel]     // 多筆商品銷售記錄
-    var totalAmount: Double
+    var totalAmount: Decimal
     var paymentMethod: PaymentMethod
     var timestamp: Date
 }
